+++ src/Physics/PhysicsEngine.h
@@ -49,8 +49,8 @@ class PHYSICS_DLL PhysicsEngine {
 	physx::PxDefaultAllocator m_Allocator;
 	physx::PxDefaultErrorCallback m_Error;
 
-	float m_Accumulator = 0.0f;
-	const float m_StepTime = 1.0f / 60.0f;
+	double m_Accumulator = 0.0f;
+	const double m_StepTime = 1.0f / 60.0f;
 
 	std::vector<physx::PxRigidActor*> m_Actors;
 	std::vector<PhysicsBody*> m_Bodies;
